requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Jinja2==3.1.3
SQLAlchemy==2.0.29
pyodbc==5.1.0
python-dotenv==1.0.1